Split test suites
- Split test suites into smoke, upgrades, features and recovery
- Add recovery partition and cos-feature tests (k3s)
- tests: respect SUT env
- Add smoke test to check we can boot into recovery

+++ packages/cos-features/cos-feature.sh
@@ -46,6 +46,10 @@ enable() {
     if [ ! -d "$OEMDIR/features" ]; then
       mkdir $OEMDIR/features
     fi
+    if [ -L "$OEMDIR/features/$i.yaml" ]; then
+      echo "Feature $i enabled already"
+      continue
+    fi
     ln -s $FEATURESDIR/$i.yaml $OEMDIR/features/$i.yaml
     echo "$i enabled"
   done
